Refactor and Overhaul
Refactoring to remove Event Hubs + Fabric Bot integration
and revise how initialization is performed.  Going forward, only a single
copy of each model will be loaded.  Model initialization is now performed
in the body of the prediction service; the prior approach to initializing
in the constructor made it more difficult to track failures, as there was
no corresponding Http Request tied to it.