[Issue Labeler] Load models on-demand
The focus of these changes is to change the approach to initialization
so that models are loaded on-demand, as needed.  Because the use has
changed to an HTTP-triggered function with low traffic, the instance seems
to be getting unloaded frequently.

With the current "initialize all models"
approach, the first request takes 60-120 seconds, on average, to return a
response  - and we often don't see additional requests come in before
the instance is unloaded.